Updating parameters to allow for non-array requests
